INDIVIDUALS ARE LISTS AGAIN
But.... convert to tuple before passing to cached functions.

This allows the individuals to be mutable AND allows for caching.

Crossover and mutation functions must also now return a list containing their offspring, rather than just the offspring itself. This applies even when there is only 1 offspring.

Offspring also verified to make sure they are valid (not all False). If an offspring individual is invalid, as random gene will be set to True.